Remove MOUNT_UNMOUNT_FILESYSTEMS from User permission
The OMA DM service requests the permission MOUNT_UNMOUNT_FILESYSTEMS. This is
excessive, and would give an attacker access to very dangerous functionality if
they compromised the OMA DM service. This permission should not be requested,
and databases should be moved from /system/etc to /data/data/come.android.omadm.
service if this causes a problem accessing these files.

bug:16711460
Change-Id: Iee307af355af02e5654e38f6615656fd59af9da9