Fix the IMEI issue, as well as the authentication issue
The imei read from the share preference is null, need call the telephony api to read
the imei directly. The authentication error is caused by missing the default case, which
will use the default server/client passwd from the dmtree.

Also add the name check with query the apn from the DB, and when apn name is  changed,
stored in SharePreference.

Bug: 17541536
Change-Id: I19bac387b4d19d61c348944c8dcd5fe92be63082